Speed up tests
We don't need to wait 2 seconds to start emitting metrics in most cases.
Instead we only need to wait long enough for Puma to finish booting and
emit the Ctrl-C line to the logs, which is what our server wrapper uses
to know when Puma is booted. The only time this might not be true is
when booting multiple workers, so I've left that at 2 seconds.

The other cleanup here is consolidating some test assertions - they were
basically doing the same thing, and effectively testing the same thing.
And with RSpec's aggregate_failures, we'll see multiple failed
assertions per test anyhow.